+++ torchmetrics/classification/auroc.py
@@ -63,12 +63,13 @@ class AUROC(Metric):
            If not ``None``, calculates standardized partial AUC over the
            range [0, max_fpr]. Should be a float between 0 and 1.
        compute_on_step:
-           Forward only calls ``update()`` and return None if this is set to False. default: True
+           Forward only calls ``update()`` and return None if this is set to False.
        dist_sync_on_step:
            Synchronize metric state across processes at each ``forward()``
            before returning the value at the step.
        process_group:
-           Specify the process group on which synchronization is called. default: None (which selects the entire world)
+           Specify the process group on which synchronization is called.
+
        dist_sync_fn:
            Callback that performs the allgather operation on the metric state. When ``None``, DDP
            will be used to perform the allgather